Update the location of models and services. Retrieve the icon from the API. Add a placeholder for charts without icon
