+++ src/ui/src/app/chart-details/chart-details-readme/chart-details-readme.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Chart } from '../../chart';
-import { ChartsService } from '../../charts.service';
+import { Chart } from '../../shared/models/chart';
+import { ChartsService } from '../../shared/services/charts.service';
 
 @Component({
   selector: 'app-chart-details-readme',
