+++ src/ui/src/app/chart-item/chart-item.component.html
@@ -1,6 +1,6 @@
 <div class="chart-item {{ fixHeight ? 'chart-item--fixHeight' : '' }}">
   <div class="chart-item-content">
-    <img alt="{{ chart.attributes.name }}'s logo" src='assets/images/k8s-icon.png'/>
+    <img class="chart-item-logo" alt="{{ chart.attributes.name }}'s logo" src="{{ getIconUrl(chart) }}"/>
     <div class="chart-item-info">
       <h2 class="chart-item-info__title" (click)="goToDetail(chart)">
         {{ chart.attributes.name }}</h2>
